api router 변경, webpack 기본로드프로그램 수정, 기타 사용자 권한관련, 환경설정 수정

+++ app/Http/Controllers/Admin/UserController.php
@@ -29,7 +29,6 @@ class UserController extends Controller {
         return view('admin.user.create', compact('roles', 'status_cd_list'));
     }
 
-
     public function store(Request $request) {
 
         $this->validate($request, [
@@ -80,9 +79,8 @@ class UserController extends Controller {
         $user = User::findorFail($id);
 
         $status_cd_list = Code::whereGroup('user_status')->get();
-
         $roles = Role::getArrayByName();
-        $userRole = $user->roles->pluck('id', 'id')->toArray();
+        $userRole = $user->roles->pluck('id', 'name')->toArray();
 
         return view('admin.user.edit', compact('user', 'roles', 'userRole', 'status_cd_list'));
     }
@@ -123,17 +121,20 @@ class UserController extends Controller {
         $user = User::findOrFail($id);
         $user->update($input);
 
-        // roles
-        DB::table('role_user')->where('user_id', $id)->delete();
-        foreach ($input['roles'] as $key => $value) {
-            $user->attachRole($value);
+        if ($user->id != 1) {
+            // roles
+            DB::table('role_user')->where('user_id', $id)->delete();
+            foreach ($input['roles'] as $key => $value) {
+                $user->attachRole($value);
+            }
+
+            // role 처리
+            DB::table('role_user')->where('user_id', $id)->delete();
+            foreach ($request->input('roles') as $key => $value) {
+                $user->attachRole($value);
+            }
         }
 
-        // role 처리
-        DB::table('role_user')->where('user_id', $id)->delete();
-        foreach ($request->input('roles') as $key => $value) {
-            $user->attachRole($value);
-        }
 
         // 아바타 변경
         if ($request->file('avatar')) {
@@ -148,12 +149,19 @@ class UserController extends Controller {
     public function destroy(Request $request, $id) {
         $user = User::findOrFail($id);
 
-        DB::table('role_user')->where('user_id', $id)->delete();
-        $user->delete();
+        if ($user->id != 1) {
 
-        return redirect()
+            DB::table('role_user')->where('user_id', $id)->delete();
+            $user->delete();
+
+            return redirect()
+                            ->route('user.index')
+                            ->with('success', trans('admin/user.destroyed'));
+        }else{
+            return redirect()
                         ->route('user.index')
-                        ->with('success', trans('admin/user.destroyed'));
+                        ->with('success', trans('admin/user.can_not_destroyed'));
+        }
     }
 
 }
